+++ src/OpenTelemetry.Instrumentation.SqlClient/CHANGELOG.md
@@ -5,7 +5,7 @@
 * Add the `db.operation.name` attribute when `CommandType` is
   `StoredProcedure` to conform to the new semantic conventions.
   This affects you if you have opted into the new conventions.
-  ([#TBD](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/TBD))
+  ([#2800](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2800))
 
 ## 1.12.0-beta.1